Replace pillowpc2001.net links.
Mike L retired from CC in 2018 and took his website down with him.
Update links which pointed to pillowpc2001.net to new locations.
His programs can be found on CC Zone. Some stuff, like the level set
list and the homepages for the CCLPx packs, moved to bitbusters.club.
The only thing that seems to be totally missing is the AVI archive.