Rewrite links section
Updated the links section to include some more current resources, and
remove some obsolete resources.

The most notable addition is bitbusters.club, which launched about a
year ago(?), and seems to be the new hub for all things CC, replacing
Mike L's site.

Discussion has moved as time has gone on. It used to be at annexcafe,
then CC Zone, and now Discord seems to be the place to be. I couldn't
bear to delete annexcafe entirely, given its enormous importance to the
history of CC (even though it shut down a few years ago), so i created a
new section for obsolete sites, and rewrote its description to provide
some historical context.

Other obsolete sites are the Yahoo! group, which is shutting down next
month, CCZone, which is reportedly on the way out, and the CC FAQ, which
hasn't been updated since 2003.